Move service mocks to right places
