+++ front/app/hooks/__mocks__/useCustomPageSlugById.ts
@@ -1,4 +1,4 @@
-import pages from '../fixtures/customPages';
+import pages from 'services/__mocks__/customPages';
 
 const customPageSlugById = pages.reduce((acc, page) => {
   acc[page.id] = `/pages/${page.attributes.slug}`;
