+++ front/app/services/__mocks__/navbarItems.ts
